requirements.txt
Babel==2.9.1
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fake-useragent==0.1.11
Flask==2.0.1
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
googlesearch-python==1.1.0
greenlet==1.1.1
gunicorn==20.1.0
idna==2.10
IMDbPY==2021.4.18
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.7.1
MarkupSafe==2.0.1
passlib==1.7.4
python-dotenv==0.19.2
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
